# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bookit API Documentation",
        "title": "Bookit API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/houses": {
            "get": {
                "description": "Returns a list of all houses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Get all houses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.HouseListItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new house. Auth required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Create a new house",
                "parameters": [
                    {
                        "description": "House data",
                        "name": "house",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.HouseCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.House"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/houses/check-slug": {
            "get": {
                "description": "Checks if a house slug is available for use",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Check house slug availability",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug to check",
                        "name": "slug",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.SlugCheckResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/houses/images/{image_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes image from database and AWS S3",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Delete house image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Image ID",
                        "name": "image_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/houses/{id}": {
            "get": {
                "description": "Returns a single house by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Get house by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.House"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a house by ID. Auth required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Delete house",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates house data. Auth required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Update house",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "House update data",
                        "name": "house",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.HouseUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.House"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/houses/{id}/images": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload multiple images for a house (max 15 images per house)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Upload house images",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "Images",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.House": {
            "type": "object",
            "properties": {
                "address_en": {
                    "type": "string"
                },
                "address_kz": {
                    "type": "string"
                },
                "address_ru": {
                    "type": "string"
                },
                "bath_qty": {
                    "type": "integer"
                },
                "bedroom_qty": {
                    "type": "integer"
                },
                "best_house": {
                    "type": "boolean"
                },
                "city_id": {
                    "type": "integer"
                },
                "comments_en": {
                    "type": "string"
                },
                "comments_kz": {
                    "type": "string"
                },
                "comments_ru": {
                    "type": "string"
                },
                "country_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description_en": {
                    "type": "string"
                },
                "description_kz": {
                    "type": "string"
                },
                "description_ru": {
                    "type": "string"
                },
                "district_en": {
                    "type": "string"
                },
                "district_kz": {
                    "type": "string"
                },
                "district_ru": {
                    "type": "string"
                },
                "guest_qty": {
                    "type": "integer"
                },
                "guests_with_pets": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Image"
                    }
                },
                "is_active": {
                    "type": "boolean"
                },
                "lat": {
                    "type": "number"
                },
                "like_count": {
                    "type": "integer"
                },
                "lng": {
                    "type": "number"
                },
                "name_en": {
                    "type": "string"
                },
                "name_kz": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "priority": {
                    "type": "integer"
                },
                "promotion": {
                    "type": "boolean"
                },
                "rooms_qty": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "type_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Image": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_label": {
                    "type": "boolean"
                },
                "mimetype": {
                    "type": "string"
                },
                "original": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "thumbnail": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation failed"
                }
            }
        },
...
        "schemas.HouseCreateRequest": {
            "description": "Request body for creating a new house",
            "type": "object",
            "required": [
                "address_en",
                "address_kz",
                "address_ru",
                "description_en",
                "description_kz",
                "description_ru",
                "name_en",
                "name_kz",
                "name_ru",
                "price",
                "type_id"
            ],
            "properties": {
                "address_en": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "123 Beach Rd"
                },
                "address_kz": {
                    "type": "string",
                    "maxLength": 255
                },
                "address_ru": {
                    "type": "string",
                    "maxLength": 255
                },
                "bath_qty": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 1
                },
                "bedroom_qty": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2
                },
                "best_house": {
                    "type": "boolean",
                    "example": false
                },
                "city_id": {
                    "type": "integer",
                    "example": 1
                },
                "country_id": {
                    "type": "integer",
                    "example": 1
                },
                "description_en": {
                    "type": "string",
                    "example": "A beautiful beach house"
                },
                "description_kz": {
                    "type": "string"
                },
                "description_ru": {
                    "type": "string"
                },
                "district_en": {
                    "type": "string",
                    "maxLength": 255
                },
                "district_kz": {
                    "type": "string",
                    "maxLength": 255
                },
                "district_ru": {
                    "type": "string",
                    "maxLength": 255
                },
                "guest_qty": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 6
                },
                "guests_with_pets": {
                    "type": "boolean",
                    "example": false
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "lat": {
                    "type": "number",
                    "example": 71.4491
                },
                "lng": {
                    "type": "number",
                    "example": 51.1694
                },
                "name_en": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Beach House"
                },
                "name_kz": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Жағажай үйі"
                },
                "name_ru": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Пляжный дом"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "+77001234567"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 50000
                },
                "priority": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "promotion": {
                    "type": "boolean",
                    "example": false
                },
                "rooms_qty": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "beach-house"
                },
                "type_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "schemas.HouseListItem": {
            "type": "object",
            "properties": {
                "address_en": {
                    "type": "string",
                    "example": "123 Beach Rd"
                },
                "address_kz": {
                    "type": "string"
                },
                "address_ru": {
                    "type": "string"
                },
                "best_house": {
                    "type": "boolean",
                    "example": false
                },
                "country_city_name_en": {
                    "type": "string",
                    "example": "Kazakhstan, Astana"
                },
                "country_city_name_kz": {
                    "type": "string",
                    "example": "Қазақстан, Астана"
                },
                "country_city_name_ru": {
                    "type": "string",
                    "example": "Казахстан, Астана"
                },
                "guests_with_pets": {
                    "type": "boolean",
                    "example": false
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.Image"
                    }
                },
                "like_count": {
                    "type": "integer",
                    "example": 5
                },
                "name_en": {
                    "type": "string",
                    "example": "Beach House"
                },
                "name_kz": {
                    "type": "string",
                    "example": "Жағажай үйі"
                },
                "name_ru": {
                    "type": "string",
                    "example": "Пляжный дом"
                },
                "owner_full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "price": {
                    "type": "integer",
                    "example": 50000
                },
                "priority": {
                    "type": "integer",
                    "example": 0
                },
                "promotion": {
                    "type": "boolean",
                    "example": false
                },
                "slug": {
                    "type": "string",
                    "example": "beach-house"
                }
            }
        },
...
        "schemas.Image": {
            "type": "object",
            "properties": {
                "house_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "mime_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "original": {
                    "type": "string",
                    "example": "https://bucket.s3.eu-central-1.amazonaws.com/houses/original/1_123.jpg"
                },
                "size": {
                    "type": "integer",
                    "example": 204800
                },
                "thumbnail": {
                    "type": "string",
                    "example": "https://bucket.s3.eu-central-1.amazonaws.com/houses/thumbnail/1_123.webp"
                }
            }
        },
...
        "schemas.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation completed"
                }
            }
        },
        "schemas.SlugCheckResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean",
                    "example": true
                },
                "slug": {
                    "type": "string",
                    "example": "beach-house"
                }
            }
        },
...
    }
}